Convert primary key from Integer to Long

+++ src/main/java/com/flowiee/dms/base/AuditEntity.java
@@ -26,7 +26,7 @@ public class AuditEntity {
 
     @Column(name = "created_by", updatable = false)
     @CreatedBy
-    protected Integer createdBy;
+    protected Long createdBy;
 
     @JsonIgnore
     @Column(name = "last_updated_at", columnDefinition = "timestamp default current_timestamp")
