+++ src/main/java/com/flowiee/dms/controller/system/AccountController.java
@@ -45,7 +45,7 @@ public class AccountController extends BaseController {
     @Operation(summary = "Find detail account")
     @GetMapping("/{accountId}")
     @PreAuthorize("@vldModuleSystem.readAccount(true)")
-    public ApiResponse<Account> findDetailAccount(@PathVariable("accountId") Integer accountId) {
+    public ApiResponse<Account> findDetailAccount(@PathVariable("accountId") Long accountId) {
         Optional<Account> account = accountService.findById(accountId);
         if (account.isEmpty()) {
             throw new ResourceNotFoundException("Account not found!", false);
@@ -73,7 +73,7 @@ public class AccountController extends BaseController {
     @Operation(summary = "Update account")
     @PutMapping(value = "/update/{accountId}")
     @PreAuthorize("@vldModuleSystem.updateAccount(true)")
-    public ApiResponse<Account> update(@RequestBody Account account, @PathVariable("accountId") Integer accountId) {
+    public ApiResponse<Account> update(@RequestBody Account account, @PathVariable("accountId") Long accountId) {
         try {
             if (accountId <= 0 || accountService.findById(accountId).isEmpty()) {
                 throw new BadRequestException();
@@ -86,7 +86,7 @@ public class AccountController extends BaseController {
 
     @PutMapping("/update-permission/{accountId}")
     @PreAuthorize("@vldModuleSystem.updateAccount(true)")
-    public ApiResponse<List<RoleModel>> updatePermission(@RequestBody String[] actions, @PathVariable("accountId") Integer accountId) {
+    public ApiResponse<List<RoleModel>> updatePermission(@RequestBody String[] actions, @PathVariable("accountId") Long accountId) {
         try {
             if (accountId <= 0 || accountService.findById(accountId).isEmpty()) {
                 throw new BadRequestException();
@@ -111,7 +111,7 @@ public class AccountController extends BaseController {
     @Operation(summary = "Delete account")
     @DeleteMapping(value = "/delete/{accountId}")
     @PreAuthorize("@vldModuleSystem.deleteAccount(true)")
-    public ApiResponse<String> deleteAccount(@PathVariable("accountId") Integer accountId) {
+    public ApiResponse<String> deleteAccount(@PathVariable("accountId") Long accountId) {
         try {
             if (accountId <= 0 ||accountService.findById(accountId).isEmpty()) {
                 throw new ResourceNotFoundException("Account not found!", false);
@@ -125,7 +125,7 @@ public class AccountController extends BaseController {
     @Operation(summary = "Find roles of account")
     @GetMapping("/{accountId}/role")
     @PreAuthorize("@vldModuleSystem.readAccount(true)")
-    public ApiResponse<List<RoleModel>> findRolesOfAccount(@PathVariable("accountId") Integer accountId) {
+    public ApiResponse<List<RoleModel>> findRolesOfAccount(@PathVariable("accountId") Long accountId) {
         try {
             return ApiResponse.ok(roleService.findAllRoleByAccountId(accountId));
         } catch (RuntimeException ex) {
