+++ src/main/java/com/flowiee/dms/controller/system/AccountControllerView.java
@@ -44,7 +44,7 @@ public class AccountControllerView extends BaseController {
 
     @GetMapping(value = "/{id}")
     @PreAuthorize("@vldModuleSystem.readAccount(true)")
-    public ModelAndView findDetailAccountById(@PathVariable("id") Integer accountId) {
+    public ModelAndView findDetailAccountById(@PathVariable("id") Long accountId) {
         if (accountId <= 0 || accountService.findById(accountId).isEmpty()) {
             throw new ResourceNotFoundException("Account not found!", true);
         }
@@ -72,7 +72,7 @@ public class AccountControllerView extends BaseController {
     @PostMapping(value = "/update/{id}")
     @PreAuthorize("@vldModuleSystem.updateAccount(true)")
     public ModelAndView update(@ModelAttribute("account") Account accountEntity,
-                               @PathVariable("id") Integer accountId,
+                               @PathVariable("id") Long accountId,
                                HttpServletRequest request) {
         Optional<Account> accOptional = accountService.findById(accountId);
         if (accOptional.isEmpty()) {
@@ -89,7 +89,7 @@ public class AccountControllerView extends BaseController {
 
     @PostMapping(value = "/delete/{id}")
     @PreAuthorize("@vldModuleSystem.deleteAccount(true)")
-    public ModelAndView deleteAccount(@PathVariable("id") Integer accountId) {
+    public ModelAndView deleteAccount(@PathVariable("id") Long accountId) {
         Optional<Account> accountOptional = accountService.findById(accountId);
         if (accountId <= 0 || accountOptional.isEmpty()) {
             throw new ResourceNotFoundException("Account not found!", true);
@@ -102,7 +102,7 @@ public class AccountControllerView extends BaseController {
 
     @PostMapping("/update-permission/{id}")
     @PreAuthorize("@vldModuleSystem.updateAccount(true)")
-    public ModelAndView updatePermission(@PathVariable("id") Integer accountId, HttpServletRequest request) {
+    public ModelAndView updatePermission(@PathVariable("id") Long accountId, HttpServletRequest request) {
         if (accountId <= 0 || accountService.findById(accountId).isEmpty()) {
             throw new ResourceNotFoundException("Account not found!", true);
         }
